// lib/api/generated/models/models-update-creator-bank-details-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UtilsCountryCode } from './utils-country-code';

export interface ModelsUpdateCreatorBankDetailsRequest {
    /**
     * Bank & Tax Details
     */
    'bank_account_name': string;
    'bank_account_number': string;
    'bank_address': string;
    'bank_name': string;
    'bank_routing_number': string;
    'tax_country': UtilsCountryCode;
    'tax_id': string;
}